+++ tracker-card/changelog.md
@@ -1,3 +1,8 @@
+## 0.1.0
+- Renamed to tracker-card
+- Added support for custom_components as well
+- Added support for multiple `check`/`update` services
+
 ## 0.0.3
 Implemented exception for `homebridge_hidden`